dawg we inside a 1000loc monolith
