+++ Content.Server/Parallax/BiomeSystem.ConfigManager.cs
@@ -1,0 +1,148 @@
+// SPDX-FileCopyrightText: 2023 Cheackraze
+// SPDX-FileCopyrightText: 2023 Checkraze
+// SPDX-FileCopyrightText: 2024 Dvir
+// SPDX-FileCopyrightText: 2024 Ed
+// SPDX-FileCopyrightText: 2024 Leon Friedrich
+// SPDX-FileCopyrightText: 2024 MilenVolf
+// SPDX-FileCopyrightText: 2024 Nemanja
+// SPDX-FileCopyrightText: 2024 Pieter-Jan Briers
+// SPDX-FileCopyrightText: 2024 Plykiya
+// SPDX-FileCopyrightText: 2024 TemporalOroboros
+// SPDX-FileCopyrightText: 2024 deltanedas
+// SPDX-FileCopyrightText: 2025 Ark
+// SPDX-FileCopyrightText: 2025 Redrover1760
+// SPDX-FileCopyrightText: 2025 metalgearsloth
+// SPDX-FileCopyrightText: 2025 starch
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+using Content.Shared.Parallax.Biomes;
+using Content.Shared.Parallax.Biomes.Layers;
+using Robust.Shared.Prototypes;
+
+namespace Content.Server.Parallax;
+
+public sealed partial class BiomeSystem
+{
+    private void InitializeConfigManager()
+    {
+        // ConfigManager methods are now part of this partial class
+    }
+
+    private void ProtoReload(PrototypesReloadedEventArgs obj)
+    {
+        if (!obj.ByType.TryGetValue(typeof(BiomeTemplatePrototype), out var reloads))
+            return;
+
+        var query = AllEntityQuery<BiomeComponent>();
+
+        while (query.MoveNext(out var uid, out var biome))
+        {
+            if (biome.Template == null || !reloads.Modified.TryGetValue(biome.Template, out var proto))
+                continue;
+
+            SetTemplate(uid, biome, (BiomeTemplatePrototype)proto);
+        }
+    }
+
+    private void SetLoadRange(float obj)
+    {
+        // Round it up
+        _loadRange = MathF.Ceiling(obj / ChunkSize) * ChunkSize;
+        _loadArea = new Box2(-_loadRange, -_loadRange, _loadRange, _loadRange);
+    }
+
+    public void SetEnabled(Entity<BiomeComponent?> ent, bool enabled = true)
+    {
+        if (!Resolve(ent, ref ent.Comp) || ent.Comp.Enabled == enabled)
+            return;
+
+        ent.Comp.Enabled = enabled;
+        Dirty(ent, ent.Comp);
+    }
+
+    public void SetSeed(EntityUid uid, BiomeComponent component, int seed, bool dirty = true)
+    {
+        component.Seed = seed;
+
+        if (dirty)
+            Dirty(uid, component);
+    }
+
+    public void ClearTemplate(EntityUid uid, BiomeComponent component, bool dirty = true)
+    {
+        component.Layers.Clear();
+        component.Template = null;
+
+        if (dirty)
+            Dirty(uid, component);
+    }
+
+    /// <summary>
+    /// Sets the <see cref="BiomeComponent.Template"/> and refreshes layers.
+    /// </summary>
+    public void SetTemplate(EntityUid uid, BiomeComponent component, BiomeTemplatePrototype template, bool dirty = true)
+    {
+        component.Layers.Clear();
+        component.Template = template.ID;
+
+        foreach (var layer in template.Layers)
+        {
+            component.Layers.Add(layer);
+        }
+
+        if (dirty)
+            Dirty(uid, component);
+    }
+
+    /// <summary>
+    /// Adds the specified layer at the specified marker if it exists.
+    /// </summary>
+    public void AddLayer(EntityUid uid, BiomeComponent component, string id, IBiomeLayer addedLayer, int seedOffset = 0)
+    {
+        for (var i = 0; i < component.Layers.Count; i++)
+        {
+            var layer = component.Layers[i];
+
+            if (layer is not BiomeDummyLayer dummy || dummy.ID != id)
+                continue;
+
+            addedLayer.Noise.SetSeed(addedLayer.Noise.GetSeed() + seedOffset);
+            component.Layers.Insert(i, addedLayer);
+            break;
+        }
+
+        Dirty(uid, component);
+    }
+
+    public void AddMarkerLayer(EntityUid uid, BiomeComponent component, string marker)
+    {
+        component.MarkerLayers.Add(marker);
+        Dirty(uid, component);
+    }
+
+    /// <summary>
+    /// Adds the specified template at the specified marker if it exists, withour overriding every layer.
+    /// </summary>
+    public void AddTemplate(EntityUid uid, BiomeComponent component, string id, BiomeTemplatePrototype template, int seedOffset = 0)
+    {
+        for (var i = 0; i < component.Layers.Count; i++)
+        {
+            var layer = component.Layers[i];
+
+            if (layer is not BiomeDummyLayer dummy || dummy.ID != id)
+                continue;
+
+            for (var j = template.Layers.Count - 1; j >= 0; j--)
+            {
+                var addedLayer = template.Layers[j];
+                addedLayer.Noise.SetSeed(addedLayer.Noise.GetSeed() + seedOffset);
+                component.Layers.Insert(i, addedLayer);
+            }
+
+            break;
+        }
+
+        Dirty(uid, component);
+    }
+}
