+++ Content.Server/Parallax/BiomeSystem.PlayerTracker.cs
@@ -1,0 +1,111 @@
+// SPDX-FileCopyrightText: 2023 Cheackraze
+// SPDX-FileCopyrightText: 2023 Checkraze
+// SPDX-FileCopyrightText: 2024 Dvir
+// SPDX-FileCopyrightText: 2024 Ed
+// SPDX-FileCopyrightText: 2024 Leon Friedrich
+// SPDX-FileCopyrightText: 2024 MilenVolf
+// SPDX-FileCopyrightText: 2024 Nemanja
+// SPDX-FileCopyrightText: 2024 Pieter-Jan Briers
+// SPDX-FileCopyrightText: 2024 Plykiya
+// SPDX-FileCopyrightText: 2024 TemporalOroboros
+// SPDX-FileCopyrightText: 2024 deltanedas
+// SPDX-FileCopyrightText: 2025 Ark
+// SPDX-FileCopyrightText: 2025 Redrover1760
+// SPDX-FileCopyrightText: 2025 metalgearsloth
+// SPDX-FileCopyrightText: 2025 starch
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+using System.Numerics;
+using Content.Shared.Parallax.Biomes;
+using Content.Shared.Parallax.Biomes.Markers;
+using Robust.Shared.Player;
+using Robust.Shared.Utility;
+using ChunkIndicesEnumerator = Robust.Shared.Map.Enumerators.ChunkIndicesEnumerator;
+
+namespace Content.Server.Parallax;
+
+public sealed partial class BiomeSystem
+{
+    private void InitializePlayerTracker()
+    {
+        // PlayerTracker methods are now part of this partial class
+    }
+
+    /// <summary>
+    /// Processes all players and viewers to determine which chunks need to be loaded.
+    /// </summary>
+    private void ProcessPlayerChunkRequests()
+    {
+        // Get chunks in range
+        foreach (var pSession in Filter.GetAllPlayers(_playerManager))
+        {
+            if (_xformQuery.TryGetComponent(pSession.AttachedEntity, out var xform) &&
+                _handledEntities.Add(pSession.AttachedEntity.Value) &&
+                 _biomeQuery.TryGetComponent(xform.MapUid, out var biome) &&
+                biome.Enabled &&
+                CanLoad(pSession.AttachedEntity.Value))
+            {
+                var worldPos = _transform.GetWorldPosition(xform);
+                AddChunksInRange(biome, worldPos);
+
+                foreach (var layer in biome.MarkerLayers)
+                {
+                    var layerProto = ProtoManager.Index(layer);
+                    AddMarkerChunksInRange(biome, worldPos, layerProto);
+                }
+            }
+
+            foreach (var viewer in pSession.ViewSubscriptions)
+            {
+                if (!_handledEntities.Add(viewer) ||
+                    !_xformQuery.TryGetComponent(viewer, out xform) ||
+                    !_biomeQuery.TryGetComponent(xform.MapUid, out biome) ||
+                    !biome.Enabled ||
+                    !CanLoad(viewer))
+                {
+                    continue;
+                }
+
+                var worldPos = _transform.GetWorldPosition(xform);
+                AddChunksInRange(biome, worldPos);
+
+                foreach (var layer in biome.MarkerLayers)
+                {
+                    var layerProto = ProtoManager.Index(layer);
+                    AddMarkerChunksInRange(biome, worldPos, layerProto);
+                }
+            }
+        }
+    }
+
+    private bool CanLoad(EntityUid uid)
+    {
+        return !_ghostQuery.HasComp(uid) || _tags.HasTag(uid, AllowBiomeLoadingTag);
+    }
+
+    private void AddChunksInRange(BiomeComponent biome, Vector2 worldPos)
+    {
+        var enumerator = new ChunkIndicesEnumerator(_loadArea.Translated(worldPos), ChunkSize);
+
+        while (enumerator.MoveNext(out var chunkOrigin))
+        {
+            _activeChunks[biome].Add(chunkOrigin.Value * ChunkSize);
+        }
+    }
+
+    private void AddMarkerChunksInRange(BiomeComponent biome, Vector2 worldPos, IBiomeMarkerLayer layer)
+    {
+        // Offset the load area so it's centralised.
+        var loadArea = new Box2(0, 0, layer.Size, layer.Size);
+        var halfLayer = new Vector2(layer.Size / 2f);
+
+        var enumerator = new ChunkIndicesEnumerator(loadArea.Translated(worldPos - halfLayer), layer.Size);
+
+        while (enumerator.MoveNext(out var chunkOrigin))
+        {
+            var lay = _markerChunks[biome].GetOrNew(layer.ID);
+            lay.Add(chunkOrigin.Value * layer.Size);
+        }
+    }
+}
